Split up big downloader code into seperate files
