+++ zypp/zyppng/media/network/downloader.cc
@@ -1,85 +1,21 @@
+/*---------------------------------------------------------------------\
+|                          ____ _   __ __ ___                          |
+|                         |__  / \ / / . \ . \                         |
+|                           / / \ V /|  _/  _/                         |
+|                          / /__ | | | | | |                           |
+|                         /_____||_| |_| |_|                           |
+|                                                                      |
+----------------------------------------------------------------------*/
 #include <zypp/zyppng/media/network/private/downloader_p.h>
-#include <zypp/zyppng/media/network/private/request_p.h>
 #include <zypp/zyppng/media/network/private/mediadebug_p.h>
 #include <zypp/zyppng/media/network/private/networkrequesterror_p.h>
 #include <zypp/zyppng/media/network/networkrequestdispatcher.h>
-#include <zypp/zyppng/media/network/request.h>
-#include <zypp-core/zyppng/base/Timer>
-#include <zypp-core/zyppng/base/EventDispatcher>
-#include <zypp-core/AutoDispose.h>
-#include <zypp/Pathname.h>
 #include <zypp/media/TransferSettings.h>
-#include <zypp/media/MetaLinkParser.h>
-#include <zypp/ByteCount.h>
-#include <zypp/base/String.h>
-#include <zypp/PathInfo.h>
 #include <zypp/media/CurlHelper.h>
+#include <zypp/media/MediaException.h>
+#include <zypp/base/String.h>
 #include <zypp/media/CredentialManager.h>
 #include <zypp/ZConfig.h>
-#include <zypp/base/Logger.h>
-#include <zypp/ZYppFactory.h>
-
-#if ENABLE_ZCHUNK_COMPRESSION
-extern "C" {
-  #include <zck.h>
-}
-#endif
-
-#include <queue>
-#include <fcntl.h>
-#include <iostream>
-#include <fstream>
-
-#define BLKSIZE		131072
-
-namespace  {
-  constexpr auto minMetalinkProbeSize = 256; //< The maximum probe size we download before we decide we really got no metalink file
-
-  bool looks_like_metalink_data( const std::vector<char> &data )
-  {
-    if ( data.empty() )
-      return false;
-
-    const char *p = data.data();
-    while (*p == ' ' || *p == '\t' || *p == '\r' || *p == '\n')
-      p++;
-
-    if (!strncasecmp(p, "<?xml", 5))
-    {
-      while (*p && *p != '>')
-        p++;
-      if (*p == '>')
-        p++;
-      while (*p == ' ' || *p == '\t' || *p == '\r' || *p == '\n')
-        p++;
-    }
-    bool ret = !strncasecmp( p, "<metalink", 9 ) ? true : false;
-    return ret;
-  }
-
-  bool looks_like_metalink_file( const zypp::Pathname &file )
-  {
-    std::unique_ptr<FILE, decltype(&fclose)> fd( fopen( file.c_str(), "r" ), &fclose );
-    if ( !fd )
-      return false;
-    return looks_like_metalink_data( zyppng::peek_data_fd( fd.get(), 0, minMetalinkProbeSize ) );
-  }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-  bool isZchunkFile ( const zypp::Pathname &file ) {
-    std::ifstream dFile( file.c_str() );
-    if ( !dFile.is_open() )
-      return false;
-
-    constexpr std::string_view magic("\0ZCK1", 5);
-
-    std::array< char, magic.size() > lead;
-    lead.fill('\0');
-    dFile.read( lead.data(), lead.size() );
-    return ( magic == std::string_view( lead.data(), lead.size()) );
-  }
-#endif
-}
 
 namespace zyppng {
 
@@ -93,1280 +29,6 @@ namespace zyppng {
 
   DownloadPrivateBase::~DownloadPrivateBase()
   { }
-
-
-  // Initial state functions, this state is used only to kickstart the statemachine
-  void InitialState::enter(){ MIL << "Entering initial state"  << std::endl; }
-
-  void InitialState::exit(){  MIL << "Leaving initial state"  << std::endl;  }
-
-  void DownloadPrivateBase::InitialState::initiate()
-  {
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-
-    if ( spec.checkExistsOnly() ) {
-      MIL_MEDIA << "Check exists only enabled" << std::endl;
-      return _sigTransitionToDlNormalFileState.emit();
-    }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-    bool deltaZck = isZchunkFile( spec.deltaFile() );
-#endif
-    if ( spec.metalinkEnabled() ) {
-#if ENABLE_ZCHUNK_COMPRESSION
-      if ( deltaZck && spec.headerSize() > 0 ) {
-        MIL_MEDIA << "We might have a zck file, detecting metalink first" << std::endl;
-        return _sigTransitionToDetectMetalinkState.emit();
-      }
-#endif
-      MIL_MEDIA << "No zchunk data available but metalink requested, going to download metalink directly." << std::endl;
-      return _sigTransitionToDlMetaLinkInfoState.emit();
-    }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-    // no Metalink, maybe we can directly download zck
-    if ( deltaZck && spec.headerSize() > 0 ) {
-      MIL_MEDIA << "No metalink but zckunk data availble trying to download ZckHead directly." << std::endl;
-      return _sigTransitionToDLZckHeaderState.emit();
-    }
-#endif
-    MIL_MEDIA << "Fallback to normal DL" << std::endl;
-    _sigTransitionToDlNormalFileState.emit();
-  }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-  std::shared_ptr<DownloadPrivateBase::DLZckHeadState> InitialState::toDLZckHeadState()
-  {
-    // we have no mirrors, the range downloader would need to fall back to using the base URL
-    return std::make_shared<DLZckHeadState>( std::vector<Url> { stateMachine()._spec.url() }, stateMachine() );
-  }
-#endif
-
-  // Metalink detection state, we query the type of file from the server, in order to use metalink support the server
-  // needs to correctly return the metalink file content type, otherwise we proceed to not downloading a metalink file
-  DetectMetalinkState::DetectMetalinkState(DownloadPrivate &parent) : SimpleState( parent ){
-    MIL_MEDIA << "Creating DetectMetalinkState" << std::endl;
-  }
-
-  void DownloadPrivateBase::DetectMetalinkState::enter()
-  {
-    _request.reset();
-    _gotMetalink = false;
-
-    auto &sm = stateMachine();
-    const auto &url = sm._spec.url();
-
-    MIL_MEDIA << "Detecting if metalink is available on " << url << std::endl;
-
-    _request = std::make_shared<Request>( ::internal::clearQueryString( url ), zypp::Pathname("/dev/null") );
-
-    _request->_originalUrl = url;
-    _request->transferSettings() = sm._spec.settings();
-    _request->transferSettings().addHeader("Accept: */*, application/metalink+xml, application/metalink4+xml");
-    _request->setOptions( _request->options() | NetworkRequest::HeadRequest );
-
-    _request->connectSignals( *this );
-    sm._requestDispatcher->enqueue( _request );
-  }
-
-  void DownloadPrivateBase::DetectMetalinkState::exit()
-  {
-    _request->disconnectSignals();
-    _request.reset();
-  }
-
-  void DownloadPrivateBase::DetectMetalinkState::onRequestStarted( NetworkRequest & )
-  {
-    stateMachine()._sigStarted.emit( *stateMachine().z_func() );
-  }
-
-  void DownloadPrivateBase::DetectMetalinkState::onRequestProgress( NetworkRequest &, off_t, off_t dlnow, off_t, off_t )
-  {
-    stateMachine()._sigAlive.emit( *stateMachine().z_func(), dlnow );
-  }
-
-  void DownloadPrivateBase::DetectMetalinkState::onRequestFinished( NetworkRequest &req, const NetworkRequestError &err )
-  {
-    auto lck = stateMachine().z_func()->shared_from_this();
-    if ( req.hasError() ) {
-      WAR << "Detecing if metalink is possible for url " << req.url() << " failed with error " << err.toString() << " falling back to download without metalink." << std::endl;
-      _error = err;
-      _gotMetalink = false;
-      return _sigFinished.emit();
-    }
-
-    std::string cType = req.contentType();
-    _gotMetalink = ( cType.find("application/metalink+xml") == 0 || cType.find("application/metalink4+xml") == 0 );
-    MIL << "Metalink detection result on url " << req.url() << " is " << _gotMetalink << std::endl;
-    _sigFinished.emit();
-  }
-
-  bool DownloadPrivateBase::DetectMetalinkState::toSimpleDownloadGuard() const
-  {
-#if ENABLE_ZCHUNK_COMPRESSION
-    return !toMetalinkGuard() && !toZckHeadDownloadGuard();
-#else
-    return !toMetalinkGuard();
-#endif
-  }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-  bool DownloadPrivateBase::DetectMetalinkState::toZckHeadDownloadGuard() const
-  {
-    return !toMetalinkGuard() && stateMachine().hasZckInfo();
-  }
-
-  std::shared_ptr<DownloadPrivateBase::DLZckHeadState> DetectMetalinkState::toDLZckHeadState()
-  {
-    // we have no mirrors, the range downloader would need to fall back to using the base URL
-    return std::make_shared<DLZckHeadState>( std::vector<Url> { stateMachine()._spec.url() }, stateMachine() );
-  }
-#endif
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::enter()
-  {
-    _request.reset();
-
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-    auto url = spec.url();
-
-    TransferSettings set = spec.settings();
-    NetworkRequestError dummyErr;
-    MirrorControl::MirrorHandle mirror;
-
-    if ( _fileMirrors.size() ) {
-      const auto res = prepareNextMirror();
-      // for Delayed or OK we can just continue here
-      if ( res != MirrorHandlingStateBase::Failed )
-        return;
-    }
-    failedToPrepare();
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::mirrorReceived( MirrorControl::MirrorPick mirror )
-  {
-    auto &sm = stateMachine();
-    auto url = sm._spec.url();
-    auto set = sm._spec.settings();
-
-    auto err = setupMirror( mirror, url, set );
-    if ( err.isError() ) {
-      WAR_MEDIA << "Setting up mirror " << mirror.second->mirrorUrl << " failed with error: " << err.toString() << "(" << err.nativeErrorString() << "), falling back to original URL." << std::endl;
-      failedToPrepare();
-    }
-    startWithMirror( mirror.second, url, set );
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::failedToPrepare()
-  {
-    startWithoutMirror();
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::startWithMirror( MirrorControl::MirrorHandle mirror, const zypp::Url &url, const TransferSettings &set )
-  {
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-
-    _request = std::make_shared<Request>( ::internal::clearQueryString(url), spec.targetPath() ) ;
-    _request->_myMirror = mirror;
-    _request->_originalUrl = url;
-
-    if ( _chksumtype  && _chksumVec ) {
-      std::shared_ptr<zypp::Digest> fileDigest = std::make_shared<zypp::Digest>();
-      if ( fileDigest->create( *_chksumtype ) )
-        // to run the checksum for the full file we need to request one big range with open end
-        _request->addRequestRange( 0, 0, fileDigest, *_chksumVec );
-    }
-    _request->transferSettings() = set;
-
-    if ( !initializeRequest( _request ) ) {
-      return failed( "Failed to initialize request" );
-    }
-
-    if ( stateMachine().previousState() && *stateMachine().previousState() != Download::InitialState ) {
-      //make sure this request will run asap
-      _request->setPriority( sm._defaultSubRequestPriority );
-    }
-
-    _request->connectSignals( *this );
-    sm._requestDispatcher->enqueue( _request );
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::startWithoutMirror()
-  {
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-
-    auto url = spec.url();
-    auto set = spec.settings();
-    auto err = sm.safeFillSettingsFromURL( url, set );
-    if ( err.isError() )
-      return failed( std::move(err) );
-    startWithMirror( nullptr, url, set );
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::exit()
-  {
-    _request->disconnectSignals();
-    _request.reset();
-  }
-
-  bool DownloadPrivateBase::BasicDownloaderStateBase::initializeRequest( std::shared_ptr<DownloadPrivateBase::Request> )
-  {
-    return true;
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::gotFinished()
-  {
-    _sigFinished.emit();
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::failed( std::string &&str )
-  {
-    failed( NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, std::move(str) ) );
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::failed( NetworkRequestError &&err )
-  {
-    _error = std::move( err );
-    _sigFailed.emit();
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::onRequestStarted( NetworkRequest & )
-  {
-    auto &sm = stateMachine();
-    if ( !sm._emittedSigStart ) {
-      sm._emittedSigStart = true;
-      stateMachine()._sigStarted.emit( *stateMachine().z_func() );
-    }
-    if ( _request->_myMirror )
-      _request->_myMirror->startTransfer();
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::handleRequestProgress( NetworkRequest &req, off_t dltotal, off_t dlnow )
-  {
-    auto &sm = stateMachine();
-    const off_t expFSize = sm._spec.expectedFileSize();
-    if ( expFSize  > 0 && expFSize < dlnow ) {
-      sm._requestDispatcher->cancel( req, NetworkRequestErrorPrivate::customError( NetworkRequestError::ExceededMaxLen ) );
-      return;
-    }
-    return sm._sigProgress.emit( *sm.z_func(), (expFSize > 0 ? expFSize : dltotal), dlnow );
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::onRequestProgress( NetworkRequest &req, off_t dltotal, off_t dlnow, off_t, off_t )
-  {
-    handleRequestProgress( req, dltotal, dlnow );
-  }
-
-  void DownloadPrivateBase::BasicDownloaderStateBase::onRequestFinished( NetworkRequest &req, const NetworkRequestError &err )
-  {
-    auto lck = stateMachine().z_func()->shared_from_this();
-    auto &sm = stateMachine();
-
-    if ( _request->_myMirror )
-      _request->_myMirror->finishTransfer( !err.isError() );
-
-    if ( req.hasError() ) {
-      // if we get authentication failure we try to recover
-      if ( sm.handleRequestAuthError( _request, err ) ) {
-        //make sure this request will run asap
-        _request->setPriority( sm._defaultSubRequestPriority );
-        sm._requestDispatcher->enqueue( _request );
-        return;
-      }
-      MIL_MEDIA << "Downloading on " << stateMachine()._spec.url() << " failed with error "<< err.toString() << " " << err.nativeErrorString() << std::endl;
-      return failed( NetworkRequestError(err) );
-    }
-
-    gotFinished();
-  }
-
-  DlMetaLinkInfoState::DlMetaLinkInfoState(DownloadPrivate &parent)
-    : BasicDownloaderStateBase( parent )
-  {
-    MIL_MEDIA << "Downloading metalink on " << parent._spec.url() << std::endl;
-  }
-
-  std::shared_ptr<DownloadPrivateBase::FinishedState> DownloadPrivateBase::DlMetaLinkInfoState::transitionToFinished()
-  {
-    MIL_MEDIA << "Downloading on " << stateMachine()._spec.url() << " transition to final state. " << std::endl;
-    return std::make_shared<FinishedState>( std::move(_error), stateMachine() );
-  }
-
-  bool DlMetaLinkInfoState::initializeRequest( std::shared_ptr<Request> r )
-  {
-    r->transferSettings().addHeader("Accept: */*, application/metalink+xml, application/metalink4+xml");
-    return BasicDownloaderStateBase::initializeRequest(r);
-  }
-
-  void DlMetaLinkInfoState::gotFinished()
-  {
-    // some proxies do not store the content type, so also look at the file to find
-    // out if we received a metalink (bnc#649925)
-    if ( !_isMetalink )
-      _isMetalink = looks_like_metalink_file( _request->targetFilePath() );
-    if ( !_isMetalink ) {
-      // Move to finished state
-      MIL << "Downloading on " << stateMachine()._spec.url() << " was successful, no metalink data. " << std::endl;
-      return BasicDownloaderStateBase::gotFinished();
-    }
-
-    // Move to Prepare Multi state
-    MIL << "Downloading on " << stateMachine()._spec.url() << " returned a Metalink " << std::endl;
-    _sigGotMetalink.emit();
-  }
-
-  void DlMetaLinkInfoState::handleRequestProgress(NetworkRequest &req, off_t dltotal, off_t dlnow)
-  {
-    auto &sm = stateMachine();
-
-    if ( !_isMetalink && dlnow < minMetalinkProbeSize ) {
-      // can't tell yet, ...
-      return sm._sigAlive.emit( *sm.z_func(), dlnow );
-    }
-
-    if ( !_isMetalink ) {
-      std::string cType = req.contentType();
-      _isMetalink = ( cType.find("application/metalink+xml") == 0 || cType.find("application/metalink4+xml") == 0 );
-    }
-
-    if ( !_isMetalink ) {
-      _isMetalink = looks_like_metalink_data( req.peekData( 0, minMetalinkProbeSize ) );
-    }
-
-    if ( _isMetalink ) {
-      // this is a metalink file change the expected filesize
-      if ( zypp::ByteCount( 2, zypp::ByteCount::MB) < static_cast<zypp::ByteCount::SizeType>( dlnow ) ) {
-        WAR << "Metalink file exceeds 2MB in filesize, aborting."<<std::endl;
-        sm._requestDispatcher->cancel( req, NetworkRequestErrorPrivate::customError( NetworkRequestError::ExceededMaxLen ) );
-        return;
-      }
-    } else {
-      // still no metalink, we assume a normal download, not perfect though
-      return BasicDownloaderStateBase::handleRequestProgress( req, dltotal, dlnow );
-    }
-  }
-
-  PrepareMultiState::PrepareMultiState(DownloadPrivate &parent) : SimpleState( parent )
-  {
-    MIL_MEDIA << "About to enter PrepareMultiState" << std::endl;
-  }
-
-  void DownloadPrivateBase::PrepareMultiState::enter( )
-  {
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-    const auto &url = spec.url();
-    const auto &targetPath = spec.targetPath();
-#if ENABLE_ZCHUNK_COMPRESSION
-    _haveZckData = (isZchunkFile( spec.deltaFile() )  && spec.headerSize() > 0);
-    DBG_MEDIA << " Upgrading request for URL: "<< url << " to multipart download , which zckunk=" << _haveZckData << std::endl;
-#else
-    DBG_MEDIA << " Upgrading request for URL: "<< url << " to multipart download , which zckunk=false" << std::endl;
-#endif
-
-
-    //we have a metalink download, lets parse it and see what we got
-    _mirrors.clear();
-
-    std::vector<zypp::media::MetalinkMirror> mirrs;
-
-    try {
-      zypp::media::MetaLinkParser parser;
-      parser.parse( targetPath );
-
-      // we only care about the metalink chunks if we have no zchunk data
-#if ENABLE_ZCHUNK_COMPRESSION
-      if ( !_haveZckData ) {
-#else
-      if ( true ) {
-#endif
-        auto bl = parser.getBlockList();
-        if ( !bl.haveBlocks() )
-          MIL_MEDIA << "Got no blocks for URL " << spec.url() << " but got filesize? " << bl.getFilesize() << std::endl;
-        if ( bl.haveBlocks() || bl.haveFilesize() )
-          _blockList = std::move(bl);
-      }
-
-      //migrate some settings from the base url to the mirror
-      mirrs = parser.getMirrors();
-      for ( auto urliter = mirrs.begin(); urliter != mirrs.end(); ++urliter ) {
-        try {
-          const std::string scheme = urliter->url.getScheme();
-          if (scheme == "http" || scheme == "https" || scheme == "ftp" || scheme == "tftp") {
-            if ( !sm._requestDispatcher->supportsProtocol( urliter->url )) {
-              urliter = mirrs.erase( urliter );
-              continue;
-            }
-            urliter->url = ::internal::propagateQueryParams( urliter->url, url );
-            _mirrors.push_back( urliter->url );
-          }
-        }
-        catch (...) {  }
-      }
-
-      if ( mirrs.empty() ) {
-        mirrs.push_back( { 0, -1, url } );
-        _mirrors.push_back( url );
-      }
-
-    } catch ( const zypp::Exception &ex ) {
-      _error = NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, zypp::str::Format("Failed to parse metalink information.(%1%)" ) % ex.asUserString() );
-      _sigFailed.emit();
-      return;
-    }
-
-    if ( mirrs.size() == 0 ) {
-      _error = NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, zypp::str::Format("Invalid metalink information.( No mirrors in metalink file)" ) );
-      _sigFailed.emit();
-      return;
-    }
-
-    //remove the metalink file
-    zypp::filesystem::unlink( targetPath );
-    _mirrorControlReadyConn = sm._mirrorControl->connect( &MirrorControl::sigNewMirrorsReady, *this, &PrepareMultiState::onMirrorsReady );
-
-    // this will emit a mirrorsReady signal once some connection tests have been done
-    sm._mirrorControl->registerMirrors( mirrs );
-  }
-
-  void DownloadPrivateBase::PrepareMultiState::onMirrorsReady()
-  {
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-    const auto &url = spec.url();
-    _mirrorControlReadyConn.disconnect();
-
-#if ENABLE_ZCHUNK_COMPRESSION
-    if ( _haveZckData  ) {
-      _sigFinished.emit();
-      return;
-    }
-#endif
-
-    // we have no zchunk data, so for a multi download we need a blocklist
-    if ( !_blockList.haveBlocks() )  {
-      //if we have no filesize we can not generate a blocklist, we need to fall back to normal download
-      if ( !_blockList.haveFilesize() ) {
-
-        //fall back to normal download but use a mirror from the mirror list
-        //otherwise we get HTTPS to HTTP redirect errors
-        _sigFallback.emit();
-        return;
-      } else {
-        //we generate a blocklist on the fly based on the filesize
-
-        MIL_MEDIA << "Generate blocklist, since there was none in the metalink file." << url  << std::endl;
-
-        off_t currOff = 0;
-        off_t filesize = _blockList.getFilesize();
-        while ( currOff <  filesize )  {
-
-          auto blksize = filesize - currOff ;
-          if ( blksize > sm._spec.preferredChunkSize() )
-            blksize = sm._spec.preferredChunkSize();
-
-          _blockList.addBlock( currOff, blksize );
-          currOff += blksize;
-        }
-
-        MIL_MEDIA << "Generated blocklist: " << std::endl << _blockList << std::endl << " End blocklist " << std::endl;
-      }
-    }
-
-    _sigFinished.emit();
-  }
-
-  std::shared_ptr<DownloadPrivateBase::DlNormalFileState> DownloadPrivateBase::PrepareMultiState::fallbackToNormalTransition()
-  {
-    MIL_MEDIA << "No blocklist and no filesize, falling back to normal download for URL " << stateMachine()._spec.url() << std::endl;
-    auto ptr = std::make_shared<DlNormalFileState>( stateMachine() );
-    ptr->_fileMirrors = std::move(_mirrors);
-
-    if ( _blockList.haveFileChecksum() ) {
-      ptr->_chksumtype = _blockList.fileChecksumType();
-      ptr->_chksumVec  = _blockList.getFileChecksum();
-    }
-
-    return ptr;
-  }
-
-  std::shared_ptr<DownloadPrivateBase::DlMetalinkState> PrepareMultiState::transitionToMetalinkDl()
-  {
-    return std::make_shared<DlMetalinkState>( std::move(_blockList), std::move(_mirrors), stateMachine() );
-  }
-
-  std::shared_ptr<DownloadPrivateBase::FinishedState> DownloadPrivateBase::PrepareMultiState::transitionToFinished()
-  {
-    return std::make_shared<FinishedState>( std::move(_error), stateMachine() );
-  }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-  std::shared_ptr<DownloadPrivateBase::DLZckHeadState> PrepareMultiState::transitionToZckHeadDl()
-  {
-    return std::make_shared<DLZckHeadState>( std::move(_mirrors), stateMachine() );
-  }
-
-  bool DownloadPrivateBase::PrepareMultiState::toZckHeadDownloadGuard() const
-  {
-    return ( stateMachine().hasZckInfo() );
-  }
-#endif
-
-  bool DownloadPrivateBase::PrepareMultiState::toMetalinkDownloadGuard() const
-  {
-#if ENABLE_ZCHUNK_COMPRESSION
-    return (!toZckHeadDownloadGuard());
-#else
-    return true;
-#endif
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::onRequestStarted( NetworkRequest & )
-  { }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::onRequestProgress( NetworkRequest &, off_t , off_t, off_t , off_t  )
-  {
-    off_t dlnowMulti = _downloadedMultiByteCount;
-    for( const auto &req : _runningRequests ) {
-      dlnowMulti += req->downloadedByteCount();
-    }
-    stateMachine()._sigProgress.emit( *stateMachine().z_func(), _fileSize, dlnowMulti );
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::onRequestFinished( NetworkRequest &req, const zyppng::NetworkRequestError &err )
-  {
-    auto lck = stateMachine().z_func()->shared_from_this();
-    auto it = std::find_if( _runningRequests.begin(), _runningRequests.end(), [ &req ]( const std::shared_ptr<Request> &r ) {
-      return ( r.get() == &req );
-    });
-    if ( it == _runningRequests.end() )
-      return;
-
-    auto reqLocked = *it;
-
-    //remove from running
-    _runningRequests.erase( it );
-
-    //feed the working URL back into the mirrors in case there are still running requests that might fail
-    // @TODO , finishing the transfer might never be called in case of cancelling the request, need a better way to track running transfers
-    if ( reqLocked->_myMirror )
-      reqLocked->_myMirror->finishTransfer( !err.isError() );
-
-    if ( err.isError() ) {
-      return handleRequestError( reqLocked, err );
-    }
-
-    _downloadedMultiByteCount += req.downloadedByteCount();
-
-    MIL_MEDIA << "Request finished "<<std::endl;
-    const auto &rngs = reqLocked->requestedRanges();
-    std::for_each( rngs.begin(), rngs.end(), []( const auto &b ){ DBG << "-> Block " << b.start << " finished." << std::endl; } );
-
-    auto restartReqWithBlock = [ this ]( std::shared_ptr<Request> &req, std::vector<Block> &&blocks ) {
-      MIL_MEDIA << "Reusing Request to download blocks:"<<std::endl;
-      if ( !addBlockRanges( req, std::move( blocks ) ) )
-        return false;
-
-      //this is not a new request, only add to queues but do not connect signals again
-      addNewRequest( req, false );
-      return true;
-    };
-
-    //check if we already have enqueued all blocks if not reuse the request
-    if ( _ranges.size() ) {
-      MIL_MEDIA << "Reusing to download blocks: "<<std::endl;
-      if ( !restartReqWithBlock( reqLocked, getNextBlocks( reqLocked->url().getScheme() ) ) ) {
-        return setFailed( "Failed to restart request with new blocks." );
-      }
-      return;
-
-    } else {
-      //if we have failed blocks, try to download them with this mirror
-      if ( !_failedRanges.empty() ) {
-
-        auto fblks = getNextFailedBlocks( reqLocked->url().getScheme() );
-        MIL_MEDIA << "Reusing to download failed blocks: "<<std::endl;
-        if ( !restartReqWithBlock( reqLocked, std::move(fblks) ) ) {
-            return setFailed( "Failed to restart request with previously failed blocks." );
-        }
-        return;
-      }
-    }
-
-    //feed the working URL back into the mirrors in case there are still running requests that might fail
-    _fileMirrors.push_back( reqLocked->_originalUrl );
-
-    // make sure downloads are running, at this point
-    ensureDownloadsRunning();
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::handleRequestError( std::shared_ptr<Request> req, const zyppng::NetworkRequestError &err )
-  {
-    bool retry = false;
-    auto &parent = stateMachine();
-
-
-    //Handle the auth errors explicitely, we need to give the user a way to put in new credentials
-    //if we get valid new credentials we can retry the request
-    if ( err.type() == NetworkRequestError::Unauthorized || err.type() == NetworkRequestError::AuthFailed ) {
-      retry = parent.handleRequestAuthError( req, err );
-    } else {
-
-      //if a error happens during a multi download we try to use another mirror to download the failed block
-      MIL << "Request failed " << req->extendedErrorString() << "(" << req->url() << ")" << std::endl;
-
-      NetworkRequestError dummyErr;
-
-      const auto &fRanges = req->failedRanges();
-      try {
-        std::transform( fRanges.begin(), fRanges.end(), std::back_inserter(_failedRanges), [ &req ]( const auto &r ){
-          Block b = std::any_cast<Block>(r.userData);;
-          b._failedWithErr = req->error();
-          DBG_MEDIA << "Adding failed block to failed blocklist: " << b.start << " " << b.len << " (" << req->error().toString() << " [" << req->error().nativeErrorString()<< "])" << std::endl;
-          return b;
-        });
-
-        // try to fill the open spot right away
-        ensureDownloadsRunning();
-        return;
-
-      } catch ( const zypp::Exception &ex ) {
-        //we just log the exception and fall back to a normal download
-        WAR << "Multipart download failed: " << ex.asString() << std::endl;
-      }
-    }
-
-    //if rety is true we just enqueue the request again, usually this means authentication was updated
-    if ( retry ) {
-      //make sure this request will run asap
-      req->setPriority( parent._defaultSubRequestPriority );
-
-      //this is not a new request, only add to queues but do not connect signals again
-      addNewRequest( req, false );
-      return;
-    }
-
-    //we do not have more mirrors left we can try
-    cancelAll ( err );
-
-    // not all hope is lost, maybe a normal download can work out?
-    // fall back to normal download
-    _sigFailed.emit();
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::ensureDownloadsRunning()
-  {
-    if ( _inEnsureDownloadsRunning )
-      return;
-
-    zypp::OnScopeExit clearFlag( [this]() {
-      _inEnsureDownloadsRunning = false;
-    });
-
-    _inEnsureDownloadsRunning = true;
-
-    //check if there is still work to do
-    while ( _ranges.size() || _failedRanges.size() ) {
-
-      // download was already finished
-      if ( _error.isError() )
-        return;
-
-      if ( _runningRequests.size() >= 10 )
-        break;
-
-      // prepareNextMirror will automatically call mirrorReceived() once there is a mirror ready
-      const auto &res = prepareNextMirror();
-      // if mirrors are delayed we stop here, once the mirrors are ready we get called again
-      if ( res == MirrorHandlingStateBase::Delayed )
-        return;
-      else if ( res == MirrorHandlingStateBase::Failed ) {
-        failedToPrepare();
-        return;
-      }
-    }
-
-    // check if we are done at this point
-    if ( _runningRequests.empty() ) {
-
-      if ( _failedRanges.size() || _ranges.size() ) {
-        setFailed( NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, "Unable to download all blocks." ) );
-        return;
-      }
-
-      // seems we were successfull , transition to finished state
-      setFinished();
-    }
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::mirrorReceived( MirrorControl::MirrorPick mirror )
-  {
-
-    auto &parent = stateMachine();
-    Url myUrl;
-    TransferSettings settings;
-
-    auto err = setupMirror( mirror, myUrl, settings );
-    if ( err.isError() ) {
-      WAR << "Failure to setup mirror " << myUrl << " with error " << err.toString() << "("<< err.nativeErrorString() << "), dropping it from the list of mirrors." << std::endl;
-      // if a mirror fails , we remove it from our list
-      _fileMirrors.erase( mirror.first );
-
-      // make sure this is retried
-      ensureDownloadsRunning();
-      return;
-    }
-
-    auto blocks = getNextBlocks( myUrl.getScheme() );
-    if ( !blocks.size() )
-      blocks = getNextFailedBlocks( myUrl.getScheme() );
-
-    if ( !blocks.size() ) {
-      // we have no blocks? should not happen but ensureDownloadsRunning will know what to do
-      ensureDownloadsRunning();
-      return;
-    }
-
-    const auto &spec = parent._spec;
-
-    std::shared_ptr<Request> req = std::make_shared<Request>( ::internal::clearQueryString( myUrl ), spec.targetPath(), NetworkRequest::WriteShared );
-    req->_myMirror = mirror.second;
-    req->_originalUrl = myUrl;
-    req->setPriority( parent._defaultSubRequestPriority );
-    req->transferSettings() = settings;
-
-    // if we download chunks we do not want to wait for too long on mirrors that have slow activity
-    // note: this sets the activity timeout, not the download timeout
-    req->transferSettings().setTimeout( 2 );
-
-    DBG_MEDIA << "Creating Request to download blocks:"<<std::endl;
-    if ( !addBlockRanges( req, std::move(blocks) ) ) {
-      setFailed( NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, "Failed to add blocks to request." ) );
-      return;
-    }
-
-    // we just use a mirror once per file, remove it from the list
-    _fileMirrors.erase( mirror.first );
-
-    addNewRequest( req );
-
-    // trigger next downloads
-    ensureDownloadsRunning();
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::failedToPrepare()
-  {
-    // it was impossible to find a new mirror, check if we still have running requests we can wait for, if not
-    // we can only fail at this point
-    if ( !_runningRequests.size() ) {
-      setFailed( NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, "No valid mirror found" ) );
-    }
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::reschedule()
-  {
-    bool triggerResched = false;
-    for ( auto &req : _runningRequests ) {
-      if ( req->state() == NetworkRequest::Pending ) {
-        triggerResched = true;
-        req->setPriority( NetworkRequest::Critical, false );
-      }
-    }
-    if ( triggerResched )
-      stateMachine()._requestDispatcher->reschedule();
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::addNewRequest(std::shared_ptr<Request> req , const bool connectSignals)
-  {
-    if ( connectSignals )
-      req->connectSignals( *this );
-
-    _runningRequests.push_back( req );
-    stateMachine()._requestDispatcher->enqueue( req );
-
-    if ( req->_myMirror )
-      req->_myMirror->startTransfer();
-  }
-
-  /**
-   * Just initialize the requests ranges from the internal blocklist
-   */
-  bool DownloadPrivateBase::RangeDownloaderBaseState::addBlockRanges ( std::shared_ptr<DownloadPrivateBase::Request> req , std::vector<Block> &&blocks ) const
-  {
-    req->resetRequestRanges();
-    for ( const auto &block : blocks ) {
-      if ( block.chksumVec && block.chksumtype.size() ) {
-        std::shared_ptr<zypp::Digest> dig = std::make_shared<zypp::Digest>();
-        if ( !dig->create( block.chksumtype ) ) {
-          WAR_MEDIA << "Trying to create Digest with chksum type " << block.chksumtype << " failed " << std::endl;
-          return false;
-        }
-
-        DBG_MEDIA << "Starting block " << block.start << " with checksum " << zypp::Digest::digestVectorToString( *block.chksumVec ) << "." << std::endl;
-        req->addRequestRange( block.start, block.len, dig, *block.chksumVec, std::any( block ), block.chksumCompareLen );
-      } else {
-        DBG_MEDIA << "Starting block " << block.start << " without checksum." << std::endl;
-        req->addRequestRange( block.start, block.len, {}, {}, std::any( block ) );
-      }
-    }
-    return true;
-  }
-
-  void zyppng::DownloadPrivateBase::RangeDownloaderBaseState::setFailed(NetworkRequestError &&err)
-  {
-    _error = std::move( err );
-    cancelAll( _error );
-    zypp::filesystem::unlink( stateMachine()._spec.targetPath() );
-    _sigFailed.emit();
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::setFailed(std::string &&reason)
-  {
-    setFailed( NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, std::move(reason) ) );
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::setFinished()
-  {
-    _error = NetworkRequestError();
-    _sigFinished.emit();
-  }
-
-  void DownloadPrivateBase::RangeDownloaderBaseState::cancelAll(const NetworkRequestError &err)
-  {
-    while( _runningRequests.size() ) {
-      auto req = _runningRequests.back();
-      req->disconnectSignals();
-      _runningRequests.pop_back();
-      stateMachine()._requestDispatcher->cancel( *req, err );
-      if ( req->_myMirror )
-        req->_myMirror->cancelTransfer();
-    }
-  }
-
-  std::vector<DownloadPrivateBase::Block> DownloadPrivateBase::RangeDownloaderBaseState::getNextBlocks( const std::string &urlScheme )
-  {
-    std::vector<DownloadPrivateBase::Block> blocks;    
-    const auto prefSize = static_cast<size_t>( stateMachine()._spec.preferredChunkSize() );
-    size_t accumulatedSize = 0;
-
-    bool canDoRandomBlocks = ( zypp::str::hasPrefixCI( urlScheme, "http") );
-    
-    std::optional<size_t> lastBlockEnd;
-    while ( _ranges.size() && accumulatedSize < prefSize ) {
-      const auto &r = _ranges.front();
-      
-      if ( !canDoRandomBlocks && lastBlockEnd ) {
-        if ( static_cast<const size_t>(r.start) != (*lastBlockEnd)+1 )
-          break;
-      }
-      
-      lastBlockEnd = r.start + r.len - 1;
-      accumulatedSize += r.len;
-      
-      blocks.push_back( std::move( _ranges.front() ) );
-      _ranges.pop_front();
-      
-    }
-    DBG_MEDIA << "Accumulated " << blocks.size() <<  " blocks with accumulated size of: " << accumulatedSize << "." << std::endl;
-    return blocks;
-  }
-
-  std::vector<DownloadPrivateBase::Block> DownloadPrivateBase::RangeDownloaderBaseState::getNextFailedBlocks( const std::string &urlScheme )
-  {
-    const auto prefSize = static_cast<size_t>( stateMachine()._spec.preferredChunkSize() );
-    // sort the failed requests by block number, this should make sure get them in offset order as well
-    _failedRanges.sort( []( const auto &a , const auto &b ){ return a.start < b.start; } );
-
-    bool canDoRandomBlocks = ( zypp::str::hasPrefixCI( urlScheme, "http") );
-
-    std::vector<DownloadPrivateBase::Block> fblks;
-    std::optional<size_t> lastBlockEnd;
-    size_t accumulatedSize = 0;
-    while ( _failedRanges.size() ) {
-      
-      const auto &block =_failedRanges.front();
-      
-      //we need to check if we have consecutive blocks because only http mirrors support random request ranges
-      if ( !canDoRandomBlocks && lastBlockEnd ) {
-        if ( static_cast<const size_t>(block.start) != (*lastBlockEnd)+1 )
-          break;
-      }
-      
-      lastBlockEnd = block.start + block.len - 1;
-      accumulatedSize += block.len;
-      
-      fblks.push_back( std::move( _failedRanges.front() ));
-      _failedRanges.pop_front();
-      
-      fblks.back()._retryCount += 1;
-
-      if ( accumulatedSize >= prefSize )
-        break;
-    }
-
-    return fblks;
-  }
-
-  DlMetalinkState::DlMetalinkState(zypp::media::MediaBlockList &&blockList, std::vector<Url> &&mirrors, DownloadPrivate &parent)
-    : RangeDownloaderBaseState( std::move(mirrors), parent )
-    , _blockList( std::move(blockList) )
-  {
-    MIL_MEDIA << "About to enter DlMetalinkState for url " << parent._spec.url() << std::endl;
-  }
-
-  void DownloadPrivateBase::DlMetalinkState::enter()
-  {
-    auto &sm = stateMachine();
-    const auto &spec = sm._spec;
-
-    //first we try to reuse blocks from the deltafile , if we have one
-    if ( !spec.deltaFile().empty() ) {
-      zypp::PathInfo dFileInfo ( spec.deltaFile() );
-      if ( dFileInfo.isFile() && dFileInfo.isR() ) {
-        FILE *f = fopen( spec.targetPath().asString().c_str(), "w+b" );
-        if ( !f ) {
-          setFailed( NetworkRequestErrorPrivate::customError( NetworkRequestError::InternalError, zypp::str::Format("Failed to open target file.(errno %1%)" ) % errno ) );
-          return;
-        }
-
-        try {
-          _blockList.reuseBlocks ( f, spec.deltaFile().asString() );
-        } catch ( ... ) { }
-
-        fclose( f );
-      }
-    }
-
-    // setup the base downloader
-    _error = {};
-    _ranges.clear();
-    _failedRanges.clear();
-    _downloadedMultiByteCount = 0;
-
-    if ( _blockList.haveFileChecksum() ) {
-      _fileChecksumType = _blockList.fileChecksumType();
-      _fileChksumVec    = _blockList.getFileChecksum();
-    }
-
-    const size_t fLen = _blockList.getFilesize();
-    if ( _fileSize > 0 ) {
-      // check if the file size as reported by zchunk is equal to the one we expect
-      if ( _fileSize != fLen ) {
-        return setFailed( NetworkRequestErrorPrivate::customError(
-          NetworkRequestError::ExceededMaxLen,
-          zypp::str::Format("Metalink file reports a different filesize than what was expected ( Meta: %1% != Exp: %2%).") % fLen % _fileSize )
-          );
-      }
-    } else {
-      _fileSize = fLen;
-    }
-
-    // remember how many bytes we need to download
-    size_t bytesToDl = 0;
-    for ( size_t i = 0; i < _blockList.numBlocks(); i++ ) {
-      const auto &mediaBlock = _blockList.getBlock( i );
-      _ranges.push_back(
-        Block{
-          .start = mediaBlock.off,
-          .len   = mediaBlock.size,
-          .chksumtype = _blockList.getChecksumType(),
-          .chksumVec  = _blockList.getChecksum ( i )
-        }
-      );
-
-      bytesToDl += mediaBlock.size;
-    }
-    // substract the length of the blocks we have to download from the overall file size
-    _downloadedMultiByteCount = fLen - bytesToDl;
-
-    ensureDownloadsRunning();
-  }
-
-  void DownloadPrivateBase::DlMetalinkState::exit()
-  {
-    cancelAll( NetworkRequestError() );
-  }
-
-  void DownloadPrivateBase::DlMetalinkState::setFinished()
-  {
-    if ( _fileChecksumType.size() && _fileChksumVec ) {
-      //TODO move this into a external application so we do not need to block on it
-      //need to check file digest
-      zypp::Digest dig;
-      dig.create( _fileChecksumType );
-
-      std::ifstream istrm( stateMachine()._spec.targetPath().asString(), std::ios::binary);
-      if ( !istrm.is_open() ) {
-        setFailed( "Failed to verify file digest (Could not open target file)." );
-        return;
-      }
-      if ( !dig.update( istrm ) ) {
-        setFailed( "Failed to verify file digest (Could not read target file)." );
-        return;
-      }
-
-      const auto &calculatedChksum = dig.digestVector();
-      if ( *_fileChksumVec != calculatedChksum ) {
-        setFailed( "Failed to verify file digest (Checksum did not match)." );
-        return;
-      }
-    }
-    RangeDownloaderBaseState::setFinished();
-  }
-
-  std::shared_ptr<DownloadPrivateBase::FinishedState> DlMetalinkState::transitionToFinished()
-  {
-    return std::make_shared<FinishedState>( std::move(_error), stateMachine() );
-  }
-
-  DlNormalFileState::DlNormalFileState(DownloadPrivate &parent) : BasicDownloaderStateBase( parent )
-  {
-    MIL_MEDIA << "About to enter DlNormalFileState for url " << parent._spec.url() << std::endl;
-  }
-
-  std::shared_ptr<DownloadPrivateBase::FinishedState> DlNormalFileState::transitionToFinished()
-  {
-    return std::make_shared<FinishedState>( std::move(_error), stateMachine() );
-  }
-
-#if ENABLE_ZCHUNK_COMPRESSION
-  DLZckHeadState::DLZckHeadState(std::vector<Url> &&mirrors, DownloadPrivate &parent)
-    : BasicDownloaderStateBase( parent )
-  {
-    _fileMirrors = std::move(mirrors);
-    MIL_MEDIA << "About to enter DlZckHeadState for url " << parent._spec.url() << std::endl;
-  }
-
-  bool DownloadPrivateBase::DLZckHeadState::initializeRequest(std::shared_ptr<DownloadPrivateBase::Request> r)
-  {
-    BasicDownloaderStateBase::initializeRequest( r );
-
-    const auto &s = stateMachine()._spec;
-    if ( s.headerSize() == 0 ) {
-      ERR_MEDIA << "Downloading the zck header was requested, but headersize is zero." << std::endl;
-      return false;
-    }
-
-    std::shared_ptr<zypp::Digest> digest;
-    NetworkRequest::CheckSumBytes sum;
-
-    const auto &headerSum = s.headerChecksum();
-    if ( headerSum ) {
-      digest = std::make_shared<zypp::Digest>();
-      if ( !digest->create( headerSum->type() ) ) {
-        ERR << "Unknown header checksum type " << headerSum->type() << std::endl;
-        return false;
-      }
-      sum = zypp::Digest::hexStringToUByteArray( headerSum->checksum() );
-    }
-
-    r->addRequestRange( 0, s.headerSize(), digest, sum );
-    return true;
-  }
-
-  void DownloadPrivateBase::DLZckHeadState::gotFinished()
-  {
-    if ( isZchunkFile( stateMachine()._spec.targetPath() ) )
-      return BasicDownloaderStateBase::gotFinished();
-    failed ( "Downloaded header is not a zchunk header");
-  }
-
-  std::shared_ptr<DownloadPrivateBase::DLZckState> DLZckHeadState::transitionToDlZckState()
-  {
-    MIL_MEDIA << "Downloaded the header of size: " << _request->downloadedByteCount() << std::endl;
-    return std::make_shared<DLZckState>( std::move(_fileMirrors), stateMachine() );
-  }
-
-  DLZckState::DLZckState(std::vector<Url> &&mirrors, DownloadPrivate &parent)
-    : RangeDownloaderBaseState( std::move(mirrors), parent )
-  {
-    MIL_MEDIA << "About to enter DLZckState for url " << parent._spec.url() << std::endl;
-  }
-
-  void DownloadPrivateBase::DLZckState::enter()
-  {
-    const auto &spec = stateMachine()._spec;
-
-    // setup the base downloader
-    _error = {};
-    _ranges.clear();
-    _failedRanges.clear();
-
-    // @TODO get this from zchunk file?
-    _fileSize = spec.expectedFileSize();
-
-    zypp::AutoFD src_fd = open( spec.deltaFile().asString().c_str(), O_RDONLY);
-    if(src_fd < 0)
-      return setFailed ( zypp::str::Format("Unable to open %1%") % spec.deltaFile() );
-
-    zypp::AutoDispose<zckCtx *> zck_src ( zck_create(), []( auto ptr ) { if ( ptr ) zck_free( &ptr ); } );
-    if( !zck_src )
-      return setFailed ( zypp::str::Format("%1%") % zck_get_error(NULL) );
-
-    if(!zck_init_read(zck_src, src_fd))
-      return setFailed ( zypp::str::Format( "Unable to open %1%: %2%") %  spec.deltaFile() % zck_get_error(zck_src) );
-
-    zypp::AutoFD target_fd = open( spec.targetPath().asString().c_str(), O_RDWR);
-    if(target_fd < 0)
-      return setFailed ( zypp::str::Format("Unable to open %1%") % spec.targetPath() );
-
-    zypp::AutoDispose<zckCtx *> zckTarget ( zck_create(), []( auto ptr ) { if ( ptr ) zck_free( &ptr ); } );
-    if( !zckTarget )
-      return setFailed ( zypp::str::Format("%1%") % zck_get_error(NULL) );
-
-    if(!zck_init_read(zckTarget, target_fd))
-      return setFailed ( zypp::str::Format( "Unable to open %1%: %2%") %  spec.targetPath() % zck_get_error(zckTarget) );
-
-    // Returns 0 for error, -1 for invalid checksum and 1 for valid checksum
-    switch ( zck_find_valid_chunks(zckTarget) ) {
-      case 0: // Returns 0 if there was a error
-        return setFailed ( zypp::str::Format( "Unable to open %1%: %2%") %  spec.targetPath() % zck_get_error(zckTarget) );
-      case 1: // getting a 1 would mean the file is already complete, basically impossible but lets handle it anyway
-        return setFinished();
-    }
-
-    const auto srcHashType = zck_get_chunk_hash_type( zckTarget );
-    const auto targetHashType = zck_get_chunk_hash_type( zckTarget );
-
-    const size_t fLen = zck_get_length( zckTarget );
-    if ( _fileSize > 0 ) {
-      // check if the file size as reported by zchunk is equal to the one we expect
-      if ( _fileSize != fLen ) {
-        return setFailed( NetworkRequestErrorPrivate::customError(
-          NetworkRequestError::ExceededMaxLen,
-          zypp::str::Format("Zchunk header reports a different filesize than what was expected ( Zck: %1% != Exp: %2%).") % fLen % _fileSize )
-        );
-      }
-    } else {
-      _fileSize = fLen;
-    }
-
-    if( srcHashType != targetHashType )
-      return setFailed ( zypp::str::Format( "ERROR: Chunk hash types don't match. Source Hash: %1% vs Target Hash: %2%")
-                        % zck_hash_name_from_type ( srcHashType )
-                        % zck_hash_name_from_type ( targetHashType ) );
-
-    if(!zck_copy_chunks( zck_src, zckTarget ))
-      return setFailed ( zypp::str::Format( "Unable to copy chunks from deltafile.") );
-
-    // will reset all chunks that are marked as failed back to missing
-    zck_reset_failed_chunks( zckTarget );
-
-
-    // we calculate what is already downloaded by substracting the block sizes we still need to download from the full file size
-    _downloadedMultiByteCount = _fileSize;
-
-    auto chunk = zck_get_first_chunk( zckTarget );
-    do {
-      // Get validity of current chunk: 1 = valid, 0 = missing, -1 = invalid
-      if ( zck_get_chunk_valid( chunk ) == 1 )
-        continue;
-
-      zypp::AutoFREE<char> zckDigest( zck_get_chunk_digest( chunk ) );
-      UByteArray chksumVec = zypp::Digest::hexStringToUByteArray( std::string_view( zckDigest.value() ) );
-      std::string chksumName;
-      std::optional<size_t> chksumCompareLen;
-
-      switch ( targetHashType ) {
-        case ZCK_HASH_SHA1: {
-          chksumName = zypp::Digest::sha1();
-          break;
-        }
-        case ZCK_HASH_SHA256: {
-          chksumName = zypp::Digest::sha256();
-          break;
-        }
-        case ZCK_HASH_SHA512: {
-          chksumName = zypp::Digest::sha512();
-          break;
-        }
-        case ZCK_HASH_SHA512_128: {
-          // defined in zchunk as
-          // SHA-512/128 (first 128 bits of SHA-512 checksum)
-          chksumName = zypp::Digest::sha512();
-          chksumCompareLen = chksumVec.size();
-          break;
-        }
-        default: {
-          return setFailed ( zypp::str::Format( "Unsupported chunk hash type: %1%.") % zck_hash_name_from_type( targetHashType ) );
-        }
-      }
-
-      const auto s = static_cast<off_t>( zck_get_chunk_start( chunk ) );
-      const auto l = static_cast<size_t>( zck_get_chunk_comp_size ( chunk ) );
-
-      MIL_MEDIA << "Downloading block " << s << " with length " << l << " checksum " << zckDigest.value() << " type " << chksumName << std::endl;
-
-      _ranges.push_back( Block{
-        .start = s,
-        .len   = l,
-        .chksumtype = chksumName,
-        .chksumVec  = std::move( chksumVec ),
-        .chksumCompareLen = std::move(chksumCompareLen)
-      } );
-
-      // substract the block length from the already downloaded bytes size
-      _downloadedMultiByteCount -= l;
-
-    } while ( (chunk = zck_get_next_chunk( chunk )) );
-
-    ensureDownloadsRunning();
-  }
-
-  void DownloadPrivateBase::DLZckState::exit()
-  {
-    cancelAll( NetworkRequestError() );
-  }
-
-  std::shared_ptr<DownloadPrivateBase::FinishedState> DLZckState::transitionToFinished()
-  {
-    return std::make_shared<FinishedState>( std::move(_error), stateMachine() );
-  }
-
-  void DownloadPrivateBase::DLZckState::setFinished()
-  {
-    const auto &spec = stateMachine()._spec;
-
-    zypp::AutoFD target_fd = open( spec.targetPath().asString().c_str(), O_RDONLY );
-    if( target_fd < 0 )
-      return setFailed ( zypp::str::Format("Unable to open %1%") % spec.targetPath() );
-
-    zypp::AutoDispose<zckCtx *> zckTarget ( zck_create(), []( auto ptr ) { if ( ptr ) zck_free( &ptr ); } );
-    if( !zckTarget )
-      return setFailed ( zypp::str::Format("%1%") % zck_get_error(nullptr) );
-
-    if(!zck_init_read(zckTarget, target_fd))
-      return setFailed ( zypp::str::Format( "Unable to open %1%: %2%") %  spec.targetPath() % zck_get_error(zckTarget) );
-
-    /* Validate the chunk and data checksums for the current file.
-     * Returns 0 for error, -1 for invalid checksum and 1 for valid checksum */
-    const auto res = zck_validate_checksums( zckTarget );
-    if ( res == 0 || res == -1 ) {
-      if( zck_is_error(nullptr) ) {
-        std::string err = zck_get_error(NULL);
-        zck_clear_error(NULL);
-        return setFailed( std::move(err) );
-      }
-      if( zck_is_error(zckTarget) )
-        return setFailed( zck_get_error(zckTarget) );
-      return setFailed( "zck_validate_checksums returned a unknown error." );
-    }
-
-    // everything is valid
-    RangeDownloaderBaseState::setFinished();
-  }
-#endif
-
-  FinishedState::FinishedState(NetworkRequestError &&error, DownloadPrivate &parent)
-    : SimpleState( parent )
-    , _error( std::move(error) )
-  {
-    MIL_MEDIA << "About to enter FinishedState for url " << parent._spec.url() << std::endl;
-  }
 
   bool DownloadPrivateBase::handleRequestAuthError( std::shared_ptr<Request> req, const zyppng::NetworkRequestError &err )
   {
@@ -1454,61 +116,6 @@ namespace zyppng {
     _sigProgressConn.disconnect();
     _sigFinishedConn.disconnect();
   }
-
-
-  DownloadPrivateBase::MirrorHandlingStateBase::MirrorHandlingStateBase( DownloadPrivate &parent )
-    : BasicState(parent)
-  { }
-
-  DownloadPrivateBase::MirrorHandlingStateBase::~MirrorHandlingStateBase()
-  {
-    _sigMirrorsReadyConn.disconnect();
-  }
-
-  DownloadPrivateBase::MirrorHandlingStateBase::PrepareResult DownloadPrivateBase::MirrorHandlingStateBase::prepareNextMirror()
-  {
-    auto &sm = stateMachine();
-    auto res = sm._mirrorControl->pickBestMirror( _fileMirrors );
-    if ( res.code == MirrorControl::PickResult::Again ) {
-      if ( !_sigMirrorsReadyConn )
-        _sigMirrorsReadyConn = sm._mirrorControl->connectFunc( &MirrorControl::sigNewMirrorsReady, [this](){
-          _sigMirrorsReadyConn.disconnect();
-          prepareNextMirror();
-        });
-      return Delayed;
-    } else if ( res.code == MirrorControl::PickResult::Unknown ) {
-      failedToPrepare();
-      return Failed;
-    }
-    mirrorReceived( res.result );
-    return Ok;
-  }
-
-  NetworkRequestError DownloadPrivateBase::MirrorHandlingStateBase::setupMirror( const MirrorControl::MirrorPick &pick, Url &url, TransferSettings &set )
-  {
-    auto &sm = stateMachine();
-    Url myUrl;
-    TransferSettings settings;
-
-    myUrl = *pick.first;
-
-    settings = sm._spec.settings();
-    //if this is a different host than the initial request, we reset username/password
-    if ( myUrl.getHost() != sm._spec.url().getHost() ) {
-      settings.setUsername( std::string() );
-      settings.setPassword( std::string() );
-      settings.setAuthType( std::string() );
-    }
-
-    NetworkRequestError err = sm.safeFillSettingsFromURL( myUrl, settings );
-    if ( err.type() != NetworkRequestError::NoError )
-      return err;
-
-    url = myUrl;
-    set = settings;
-    return err;
-  }
-
 
   DownloadPrivate::DownloadPrivate(Downloader &parent, std::shared_ptr<NetworkRequestDispatcher> requestDispatcher, std::shared_ptr<MirrorControl> mirrors, DownloadSpec &&spec, Download &p)
     : DownloadPrivateBase( parent, std::move(requestDispatcher), std::move(mirrors), std::move(spec), p )
@@ -1617,7 +224,7 @@ namespace zyppng {
   NetworkRequestError Download::lastRequestError() const
   {
     if ( state() == Finished ) {
-      return d_func()->state<DownloadPrivateBase::FinishedState>()->_error;
+      return d_func()->state<FinishedState>()->_error;
     }
     return NetworkRequestError();
   }
