+++ zypp/zyppng/media/network/mirrorcontrol.cc
@@ -1,3 +1,11 @@
+/*---------------------------------------------------------------------\
+|                          ____ _   __ __ ___                          |
+|                         |__  / \ / / . \ . \                         |
+|                           / / \ V /|  _/  _/                         |
+|                          / /__ | | | | | |                           |
+|                         /_____||_| |_| |_|                           |
+|                                                                      |
+----------------------------------------------------------------------*/
 #include "private/mirrorcontrol_p.h"
 #include "private/mediadebug_p.h"
 #include <zypp-core/zyppng/base/EventDispatcher>
