+++ zypp/zyppng/media/network/private/mediadebug_p.h
@@ -1,3 +1,16 @@
+/*---------------------------------------------------------------------\
+|                          ____ _   __ __ ___                          |
+|                         |__  / \ / / . \ . \                         |
+|                           / / \ V /|  _/  _/                         |
+|                          / /__ | | | | | |                           |
+|                         /_____||_| |_| |_|                           |
+|                                                                      |
+----------------------------------------------------------------------/
+*
+* This file contains private API, this might break at any time between releases.
+* You have been warned!
+*
+*/
 #ifndef ZYPP_NG_MEDIADEBUG_H_INCLUDED
 #define ZYPP_NG_MEDIADEBUG_H_INCLUDED
